+++ src/utility-types.ts
@@ -1,7 +1,7 @@
 import type { Worker } from 'cluster'
 import type { MessagePort } from 'worker_threads'
+import type { KillBehavior } from './worker/worker-options'
 
-export type KillBehavior = 'HARD' | 'SOFT'
 /**
  * Make all properties in T non-readonly
  */
@@ -43,7 +43,7 @@ export interface MessageValue<
   /**
    * Kill code.
    */
-  readonly kill?: KillBehavior | number
+  readonly kill?: KillBehavior | 1
   /**
    * Error.
    */